+++ modules/adtargetBidAdapter.js
@@ -4,60 +4,23 @@ import {BANNER, VIDEO} from '../src/mediaTypes.js';
 import {config} from '../src/config.js';
 import {find} from '../src/polyfill.js';
 import {chunk} from '../libraries/chunk/chunk.js';
+import {
+  createTag,
+  getUserSyncs,
+  isBidRequestValid,
+  supportedMediaTypes
+} from '../libraries/adtelligentUtils/adtelligentUtils.js';
 
 const ENDPOINT = 'https://ghb.console.adtarget.com.tr/v2/auction/';
 const BIDDER_CODE = 'adtarget';
 const DISPLAY = 'display';
-const syncsCache = {};
 
 export const spec = {
   code: BIDDER_CODE,
   gvlid: 779,
-  supportedMediaTypes: [VIDEO, BANNER],
-  isBidRequestValid: function (bid) {
-    return !!deepAccess(bid, 'params.aid');
-  },
-  getUserSyncs: function (syncOptions, serverResponses) {
-    const syncs = [];
-
-    function addSyncs(bid) {
-      const uris = bid.cookieURLs;
-      const types = bid.cookieURLSTypes || [];
-
-      if (Array.isArray(uris)) {
-        uris.forEach((uri, i) => {
-          const type = types[i] || 'image';
-
-          if ((!syncOptions.pixelEnabled && type === 'image') ||
-            (!syncOptions.iframeEnabled && type === 'iframe') ||
-            syncsCache[uri]) {
-            return;
-          }
-
-          syncsCache[uri] = true;
-          syncs.push({
-            type: type,
-            url: uri
-          })
-        })
-      }
-    }
-
-    if (syncOptions.pixelEnabled || syncOptions.iframeEnabled) {
-      isArray(serverResponses) && serverResponses.forEach((response) => {
-        if (response.body) {
-          if (isArray(response.body)) {
-            response.body.forEach(b => {
-              addSyncs(b);
-            })
-          } else {
-            addSyncs(response.body)
-          }
-        }
-      })
-    }
-    return syncs;
-  },
+  supportedMediaTypes,
+  isBidRequestValid,
+  getUserSyncs,
 
   buildRequests: function (bidRequests, adapterRequest) {
     const adapterSettings = config.getConfig(adapterRequest.bidderCode)
@@ -118,26 +81,7 @@ function parseResponse(serverResponse, adapterRequest) {
 }
 
 function bidToTag(bidRequests, adapterRequest) {
-  const tag = {
-    // TODO: is 'page' the right value here?
-    Domain: deepAccess(adapterRequest, 'refererInfo.page')
-  };
-  if (config.getConfig('coppa') === true) {
-    tag.Coppa = 1;
-  }
-  if (deepAccess(adapterRequest, 'gdprConsent.gdprApplies')) {
-    tag.GDPR = 1;
-    tag.GDPRConsent = deepAccess(adapterRequest, 'gdprConsent.consentString');
-  }
-  if (deepAccess(adapterRequest, 'uspConsent')) {
-    tag.USP = deepAccess(adapterRequest, 'uspConsent');
-  }
-  if (deepAccess(bidRequests[0], 'schain')) {
-    tag.Schain = deepAccess(bidRequests[0], 'schain');
-  }
-  if (deepAccess(bidRequests[0], 'userId')) {
-    tag.UserIds = deepAccess(bidRequests[0], 'userId');
-  }
+  const tag = createTag(bidRequests, adapterRequest);
 
   const bids = [];
