+++ modules/adtelligentBidAdapter.js
@@ -5,6 +5,12 @@ import {config} from '../src/config.js';
 import {Renderer} from '../src/Renderer.js';
 import {find} from '../src/polyfill.js';
 import {chunk} from '../libraries/chunk/chunk.js';
+import {
+  createTag, getUserSyncs,
+  isBidRequestValid,
+  supportedMediaTypes
+} from '../libraries/adtelligentUtils/adtelligentUtils.js';
+
 
 /**
  * @typedef {import('../src/adapters/bidderFactory.js').Bid} Bid
@@ -37,7 +43,6 @@ const OUTSTREAM_SRC = 'https://player.adtelligent.com/outstream-unit/2.01/outstr
 const BIDDER_CODE = 'adtelligent';
 const OUTSTREAM = 'outstream';
 const DISPLAY = 'display';
-const syncsCache = {};
 
 export const spec = {
   code: BIDDER_CODE,
@@ -51,51 +56,9 @@ export const spec = {
     'indicue',
     'stellormedia'
   ],
-  supportedMediaTypes: [VIDEO, BANNER],
-  isBidRequestValid: function (bid) {
-    return !!deepAccess(bid, 'params.aid');
-  },
-  getUserSyncs: function (syncOptions, serverResponses) {
-    const syncs = [];
-
-    function addSyncs(bid) {
-      const uris = bid.cookieURLs;
-      const types = bid.cookieURLSTypes || [];
-
-      if (Array.isArray(uris)) {
-        uris.forEach((uri, i) => {
-          const type = types[i] || 'image';
-
-          if ((!syncOptions.pixelEnabled && type === 'image') ||
-            (!syncOptions.iframeEnabled && type === 'iframe') ||
-            syncsCache[uri]) {
-            return;
-          }
-
-          syncsCache[uri] = true;
-          syncs.push({
-            type: type,
-            url: uri
-          })
-        })
-      }
-    }
-
-    if (syncOptions.pixelEnabled || syncOptions.iframeEnabled) {
-      isArray(serverResponses) && serverResponses.forEach((response) => {
-        if (response.body) {
-          if (isArray(response.body)) {
-            response.body.forEach(b => {
-              addSyncs(b);
-            })
-          } else {
-            addSyncs(response.body)
-          }
-        }
-      })
-    }
-    return syncs;
-  },
+  supportedMediaTypes,
+  isBidRequestValid,
+  getUserSyncs,
   /**
    * Make a server request from the list of BidRequests
    * @param bidRequests
@@ -164,29 +127,7 @@ function parseRTBResponse(serverResponse, adapterRequest) {
 
 function bidToTag(bidRequests, adapterRequest) {
   // start publisher env
-  const tag = {
-    // TODO: is 'page' the right value here?
-    Domain: deepAccess(adapterRequest, 'refererInfo.page')
-  };
-  if (config.getConfig('coppa') === true) {
-    tag.Coppa = 1;
-  }
-  if (deepAccess(adapterRequest, 'gdprConsent.gdprApplies')) {
-    tag.GDPR = 1;
-    tag.GDPRConsent = deepAccess(adapterRequest, 'gdprConsent.consentString');
-  }
-  if (deepAccess(adapterRequest, 'uspConsent')) {
-    tag.USP = deepAccess(adapterRequest, 'uspConsent');
-  }
-  if (deepAccess(bidRequests[0], 'schain')) {
-    tag.Schain = deepAccess(bidRequests[0], 'schain');
-  }
-  if (deepAccess(bidRequests[0], 'userId')) {
-    tag.UserIds = deepAccess(bidRequests[0], 'userId');
-  }
-  if (deepAccess(bidRequests[0], 'userIdAsEids')) {
-    tag.UserEids = deepAccess(bidRequests[0], 'userIdAsEids');
-  }
+  const tag = createTag(bidRequests, adapterRequest);
   if (window.adtDmp && window.adtDmp.ready) {
     tag.DMPId = window.adtDmp.getUID();
   }
